backend/requirements/base.txt
# Base requirements for NextCRM (June 2025)
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-extensions==3.2.3
Pillow==10.4.0
pytz==2024.1

# Database
psycopg==3.2.1

# Redis/Caching
redis==5.0.8
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
cryptography==42.0.8

# API Documentation
drf-spectacular==0.27.2

# Filtering
django-filter==24.2

# Environment Variables
python-decouple==3.8

# Date/Time
python-dateutil==2.9.0

# Utilities
requests==2.32.3

# Additional dependencies
setuptools==70.3.0